backend/requirements.txt
django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
drf-yasg==1.21.7
python-dateutil==2.8.2
pandas==2.2.0
Pillow==10.2.0
django-filter==23.5
celery==5.3.6
django-environ==0.11.2
django-encrypted-model-fields==0.6.5
django-celery-beat==2.5.0
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0 